+++ src/assembler.rs
@@ -492,7 +492,7 @@ fn parse_parts(source: &str) -> Result<(Module, Vec<FieldFixup>), Fault> {
                         ),
                         "sizeof" | "alignof" | "heap.alloc" | "ptr.null" | "ptr.cast"
                         | "ptr.fromint" | "ldobj" | "stobj" | "cpobj" | "initobj"
-                        | "value.pack" | "value.is" | "value.unpack" | "ldtoken"
+                        | "value.pack" | "value.is" | "value.unpack" | "ldtoken" | "castclass"
                         | "interface.borrow" | "newarr" | "array.alloc" | "array.create"
                         | "ldelem" | "stelem" | "ldelema" => Some(
                             serde_json::to_value(parse_type(rest)?)